use rename() instead of link()
so that a previous file would be overwritten (if the key is being overwritten as well)

also check the return value and output an error message if rename() fails

+++ src/storage_fs.c
@@ -124,6 +124,8 @@ st_store(void *key, size_t klen, void *value, size_t vlen, void *priv)
     char tmpdir[tmpdir_len];
     char *tmp_intermediate = NULL;
     char *intermediate_dir = NULL;
+    int renamed = 0;
+
     snprintf(tmpdir, tmpdir_len, "%s/%s", storage->tmp, dname);
     char *tmppath = st_fs_filename(tmpdir, key, klen, &tmp_intermediate);
     int fd = open(tmppath, O_WRONLY|O_TRUNC|O_CREAT, S_IRWXU|S_IRWXG|S_IRWXO);
@@ -146,13 +148,20 @@ st_store(void *key, size_t klen, void *value, size_t vlen, void *priv)
                                             key,
                                             klen,
                                             &intermediate_dir);
-            ret = link(tmppath, fullpath);
+            ret = rename(tmppath, fullpath);
+            if (ret == 0)
+                renamed = 1;
+            else
+                SHC_ERROR("Can't store data on file %s : %s\n",
+                           fullpath, strerror(errno));
+
             free(fullpath);
         }
         close(fd);
     }
 
-    unlink(tmppath);
+    if (!renamed)
+	    unlink(tmppath);
     rmdir(tmp_intermediate);
     rmdir(tmpdir);
     free(tmppath);
